% Auto-generated: do not edit by hand
\name{multiTextInput}

\alias{multiTextInput}

\title{MultiTextInput component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
multiTextInput(id=NULL, label=NULL, placeholder=NULL, value=NULL,
values=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{label}{Character. A label that will be printed when this component is rendered.}

\item{placeholder}{Character. }

\item{value}{Character. The value displayed in the input.}

\item{values}{List of characters. The values validated by the input.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

